requirements.txt
nre-darwin-py==0.4.0
openmeteo-requests==1.3.0
boto3==1.34.125
pandas==2.2.3
requests-cache==1.2.1
retry-requests==2.0.0
certifi~=2024.8.30
pydantic-settings~=2.5.2
sal==1.2.5
PyJWT~=2.9.0
openai~=1.51.2
langchain-core~=0.3.10
requests~=2.32.3
langchain-openai~=0.2.2
langgraph~=0.2.35
pillow~=11.0.0
